flaskm/requirements/common.txt
alembic==0.8.4
blinker==1.4
dominate==2.1.16
Flask==0.10.1
Flask-Bootstrap==3.3.5.7
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-Migrate==1.7.0
Flask-Moment==0.5.1
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.12
itsdangerous==0.24
Jinja2==2.8
Mako==1.0.3
MarkupSafe==0.23
python-editor==0.5
SQLAlchemy==1.0.12
visitor==0.1.2
Werkzeug==0.11.4
wheel==0.24.0
WTForms==2.1
Flask-PageDown==0.2.1
Markdown==2.6.6
bleach==1.4.2